pyproject.toml
[project]
name = "mcp-compute-scout"
version = "0.1.0"
description = "MCP server for checking compute and GPU availability across servers"
readme = "README.md"
requires-python = ">=3.8"
authors = [{ name = "matsengrp" }]
dependencies = [
    "mcp>=1.0.0",
    "pyyaml>=6.0",
    "tabulate>=0.9.0",
]

[project.scripts]
mcp-compute-scout = "mcp_compute_scout.__main__:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/mcp_compute_scout"]